# swagger.yaml
openapi: 3.0.0
info:
  title: Zendvo Auth API
  version: 1.0.0
  description: API for user registration and authentication
paths:
...
  /api/auth/login:
    post:
      summary: User login
      description: Returns access and refresh tokens for valid credentials.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: "Password123!"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      refresh_token:
                        type: string
        "401":
          description: Invalid email or password
        "429":
          description: Too many failed attempts from the same IP
        "500":
          description: Internal server error
...
  /api/auth/me:
    get:
      summary: Get current user
      description: Returns the authenticated user's profile and verification status.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User returned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
                        format: email
                      name:
                        type: string
                        nullable: true
                      role:
                        type: string
                      status:
                        type: string
                      createdAt:
                        type: string
                        format: date-time
                      lastLogin:
                        type: string
                        format: date-time
                        nullable: true
                      email_verified:
                        type: boolean
        "401":
          description: Unauthorized
        "404":
          description: User not found
        "500":
          description: Internal server error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT